+++ app/auth/sing-up/page.tsx
@@ -1,5 +1,5 @@
-import { SignUpForm } from '@/src/features/auth/sign-up/ui/SignUpForm';
+import SignUpPage from '@/src/features/auth/sign-up/ui/SignUpPage';
 
 export default function Page() {
-  return <SignUpForm />;
+  return <SignUpPage />;
 }
