+++ src/features/auth/sign-up/ui/SignInPrompt/SignInPrompt.module.scss
@@ -1,0 +1,7 @@
+.signInSection {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 8px;
+  margin-top: 18px;
+}
