+++ src/features/auth/sign-up/ui/SignUpForm/SignUpForm.module.scss
@@ -1,0 +1,30 @@
+.form {
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+  width: 100%;
+}
+
+.specialGap {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+}
+
+.customInput {
+  background-color: var(--color-dark-500);
+}
+
+.link {
+  font-family: "Inter", sans-serif;
+  font-weight: 400;
+  font-size: 12px;
+  text-decoration: underline;
+  text-decoration-skip-ink: none;
+  white-space: nowrap;
+
+  &:hover {
+    color: var(--color-accent-500);
+    text-decoration: underline;
+  }
+}
