+++ src/features/auth/sign-up/ui/SignUpPage.module.scss
@@ -1,0 +1,11 @@
+.card {
+  padding: 24px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.title {
+  margin-bottom: 13px;
+}
+
