--- src/pages/auth/SignUpPage.module.scss
@@ -1,90 +1,0 @@
-.pageWrapper {
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  padding: 24px;
-}
-
-.container {
-  width: 100%;
-  max-width: 378px;
-  margin: 0 auto;
-}
-
-.card {
-  padding: 24px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.title {
-  font-family: var(--font-family-primary);
-  font-weight: 700;
-  font-size: 20px;
-  line-height: 36px;
-  text-align: center;
-  color: var(--color-light-100);
-  margin: -1px 0 13px 0;
-  width: 100%;
-}
-
-.oauthSection {
-  margin-bottom: 24px;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-}
-
-.oauthContainer {
-  display: flex;
-  justify-content: center;
-  gap: 60px;
-}
-
-.oauthButton {
-  width: 48px;
-  height: 48px;
-  border-radius: 50%;
-  border: 1px solid var(--color-dark-300);
-  background: transparent;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  transition: all 0.2s ease;
-
-  &:hover {
-    background-color: var(--color-dark-400);
-  }
-}
-
-.oauthIcon {
-  width: 24px;
-  height: 24px;
-  color: var(--color-light-100);
-}
-
-.formContainer {
-  width: 100%;
-  padding: 0 24px; // 0 сверху/снизу, 24px по бокам
-}
-
-.signInSection {
-  text-align: center;
-  width: 100%;
-  margin-top: 24px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 12px;
-}
-
-.signInText {
-  font-size: 14px;
-  color: var(--color-light-900);
-}
-
-
